implement a case matching necessary env vars
It is necessary to handle the cases in which the plugin has all necessary environment variables to setup the `repo_server_path` on the startup, or if it only has some of them or even if it hasn't the necessary ones to configure it.

In case the environment variables are not available, it returns the default value for the `repo_server_path`.